It is doubtless that global change has become one the challenges, which encompasses a wide range of human life, including social and economical aspects of human life. Research has indicated that climate change will continue affecting the world as long as proper measures are not taken to protect the environment.
In this line of thought, human activities have been widely blamed for escalating effects of climate change around the world (Hillel & Rosenzweig 2010). Only time will tell whether taming climate change is possible or not.

In this regard, this assessment covers the impact of climate change in our lives today even as world leaders burn midnight oil to develop strategies, aimed at taming the scourge. This proposal topic has an array of benefits, especially in understanding the fatal nature of climate change.

It will mainly focus on the effects of climate change and make proposals on how to counteract the effects of climate together some of the preventive measures being considered by international leaders.

Through literature review, this project will compare different views as argued by different authors in order to synthesize the issue with varying view points. This will be crucial in capturing the main objective of the projects, which revolves around the analysis of the effects of climate change in the world today.

Climate change
How is climate change defined? Although different environmental experts tend to have different definitions, the Australian Government defines climate change as the weather pattern observed for several years. These changes are mainly caused by human activities, which negatively impact the environment.

With reference to the Intergovernmental Panel on Climate Change (IPCC) report released in the year 2007, climate change is no longer a myth, but a reality, whose impact has continually escalated from 1950s, mainly due to rising levels of greenhouse gases into the atmosphere.

This implies that human activities have significantly contributed this environmental scourge, which continues to affect most parts of the world. The IPCC report was a representation of the world view on climate change, collected from various scientific journals published around the world (Australian Government 2012).

The Australian Department of Climate Change and Energy Efficiency affirms that there is enough evidence to support the fact that the climate system of the earth has continuously been warming. Some of the observations made include the rising level of air in the world and high ocean temperatures. Others are the rising sea level, constant melting of snow and ice in most parts of the world.

One important fact to note about climate change is that it involves the rising temperatures of the climate system holistically, including all the oceans, atmosphere and the cryosphere. These findings concluded that the climate system is in a heating mode.

Even as we review other people’s work, it is important to note that climate change is more than mere global warming as perceived by most people. From scientific revelation, the climate will be varied broadly especially if the warming continues uncontrollably (Australian Government 2012). As a result, the world is likely to experience irregular rainfall patterns, occurrence of severe climatic events like heavy currents and droughts among others.

Climate change impacts
The impact of climate change has been felt in every part of the world. According to United Nations Framework Convention on Climate Change (UNFCCC), Asia, Africa and Latin America are among the regions of the world, which have severely been affected by the scourge. In a 2010 survey carried out by Climate Change Secretariat, Africa is under the pressure of climate change and remains vulnerable to these effects.

Unlike most parts of the world, Africa experiences varying climatic changes. Common occurrences in Africa are severe droughts and floods, which have had negative implications on the continent’s economy (UNFCCC 2010).

The two events are widely known to predispose famine and overall interference with the socio-well being of the society. According to the UNFCCC’s analysis, close to a third of Africa’s population inhabit drought-prone regions, while more than two million remain vulnerable to drought every year (UNFCCC 2010).

In understanding the implication of climate change in Africa, the survey found out that the issue of climate change is intertwined with several factors, which contribute to its escalation across the continent.

Some of these factors include poverty, weak institutions, illiteracy, lack of information and technology, limited infrastructure, poor accessibility to resources, poor management and conflicts. In addition, there is widespread exploitation of land, which remains a major threat to the climate.

Due to pressure on farming land, most farmers exert pressure through over-cultivation and deforestation. In addition, other factors like dunes and storms continue posing more negative threats to the environment and human beings (UNFCCC 2010).

As a result of these events, the continent experiences drought and overall scarcity of water. Due to this emerging trend, Africa is likely to face shortage of rainfall and overall scarcity of water. With Africa having several trans-boundary river basins, the continent is likely to experience conflicts over these basins. Another important aspect captured in the report is agriculture (UNFCCC 2010).

Since most subsistence farmers in Africa depend on rainfall and irrigation, the sector has been affected by insufficient supply in most Sub-Saharan regions. Besides this, UNFCCC notes that climate change has resulted into loss of agricultural land and a drop in subsistence crop production. With a good percentage of the population under the threat of starvation, climate change has undoubtedly led to escalation of insufficient food supply.

It is amazing to note that climate change has also contributed to the spread of some diseases like malaria, tuberculosis and diarrhea in most parts of Africa. As stated by the UNFCCC, there has been a shift in the distribution of disease vectors.

For instance, migration of mosquitoes to regions of higher altitude is likely to expose people in such regions to the risk of contracting malaria (UNFCCC 2010). Additionally, climate change is likely to result into negative impact on African ecosystems and habitats, which are already threatened by these changes. Due to reduced habitat and changing climatic conditions some species are likely to move to more tolerable regions.

In this line of though Robert Watson, Marufu Zinyowera and Richard Moss found out that climate change can have severe effects on human health. In a research carried out in 1998, the three reiterated that human health may be affected as a result of heat-stress mortality, urban air pollution and vector-borne diseases, which could be favored as a result of change in temperature or rainfall in a given ecosystem (Watson, Zinyowera & Moss 1998, p. 7).

Additionally, Watson, Zinyowera and Moss argued that these effects are commonly felt in developing countries, where lives are lost, communities affected and the cost in medical care rises due to high prevalence of some health complications.

With regard to the impact of climate change on biodiversity, Watson, Zinyowera and Moss, agree with UNFCCC’s findings. In their 1998 survey, the three argued that all ecosystems play a fundamental role in the society (Watson, Zinyowera & Moss 1998).

For instance, they are a source of goods and services to any